# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Aakash Enterprises - Cold Drinks Inventory & Sales Management API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface MessageResponse {
  message: string;
}

export interface Product {
  id: number;
  name: string;
  brand: string;
  size: string;
  purchasePrice: number;
  sellingPrice: number;
  stock: number;
  supplier: string;
  createdAt: string;
}

export interface CreateProductInput {
  name: string;
  brand: string;
  size: string;
  purchasePrice: number;
  sellingPrice: number;
  stock: number;
  supplier: string;
}

export interface Customer {
  id: number;
  name: string;
  phone: string;
  shopName: string;
  address: string;
  notes: string;
  createdAt: string;
}

export interface CreateCustomerInput {
  name: string;
  phone: string;
  shopName: string;
  address: string;
  notes: string;
}

export interface SaleItem {
  productId: number;
  quantity: number;
  price: number;
}

export interface CreateSaleInput {
  customerId?: number | null;
  items: SaleItem[];
  notes?: string | null;
}

export interface SaleItemWithProduct {
  id: number;
  productId: number;
  productName: string;
  quantity: number;
  price: number;
  total: number;
}

export interface SaleWithDetails {
  id: number;
  customerId?: number | null;
  customerName?: string | null;
  total: number;
  notes?: string | null;
  date: string;
  items: SaleItemWithProduct[];
}

export interface LowStockProduct {
  id: number;
  name: string;
  brand: string;
  stock: number;
}

export interface StarProduct {
  id: number;
  name: string;
  brand: string;
  quantitySold: number;
  revenue: number;
}

export interface DashboardStats {
  todaySales: number;
  todayDrinksSold: number;
  monthlySales: number;
  stockRemaining: number;
  lowStockProducts: LowStockProduct[];
  starProduct?: StarProduct | null;
}

export interface DailySalesData {
  date: string;
  revenue: number;
  quantitySold: number;
}

export interface MonthlySalesData {
  month: string;
  revenue: number;
  quantitySold: number;
}

export interface TopProduct {
  id: number;
  name: string;
  brand: string;
  size: string;
  quantitySold: number;
  revenue: number;
}

export interface ProductProfitMargin {
  id: number;
  name: string;
  brand: string;
  size: string;
  purchasePrice: number;
  sellingPrice: number;
  profitPerUnit: number;
  profitMarginPercent: number;
  totalUnitsSold: number;
  totalRevenue: number;
  totalProfit: number;
  stock: number;
}

export type GetProductsParams = {
  search?: string;
};

export type GetCustomersParams = {
  search?: string;
};

export type GetSalesParams = {
  startDate?: string;
  endDate?: string;
  customerId?: number;
};