# lib/api-zod/src/generated/types/createSaleInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Aakash Enterprises - Cold Drinks Inventory & Sales Management API
 * OpenAPI spec version: 0.1.0
 */
import type { SaleItem } from "./saleItem";

export interface CreateSaleInput {
  customerId?: number | null;
  items: SaleItem[];
  notes?: string | null;
}