# lib/api-zod/src/generated/types/customer.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Aakash Enterprises - Cold Drinks Inventory & Sales Management API
 * OpenAPI spec version: 0.1.0
 */

export interface Customer {
  id: number;
  name: string;
  phone: string;
  shopName: string;
  address: string;
  notes: string;
  createdAt: string;
}