# lib/api-zod/src/generated/types/lowStockProduct.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Aakash Enterprises - Cold Drinks Inventory & Sales Management API
 * OpenAPI spec version: 0.1.0
 */

export interface LowStockProduct {
  id: number;
  name: string;
  brand: string;
  stock: number;
}