# lib/api-zod/src/generated/types/saleWithDetails.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Aakash Enterprises - Cold Drinks Inventory & Sales Management API
 * OpenAPI spec version: 0.1.0
 */
import type { SaleItemWithProduct } from "./saleItemWithProduct";

export interface SaleWithDetails {
  id: number;
  customerId?: number | null;
  customerName?: string | null;
  total: number;
  notes?: string | null;
  date: string;
  items: SaleItemWithProduct[];
}